Automatically generate the audit events table
Ensure that we have a complete table of audit events in the docs by
generating the table automatically from the source.

This change assumes that all AuditEvent implementations include a
`Metadata` field with the value type `apievents.Metadata`. It uses the
ast and astutil packages to parse all Go files in the
gravitational/teleport source, then traverse the resulting ASTs twice:

- Build a map of Metadata types based on all assignments of an
  `apievents.Metadata` to a `Metadata` field
- Find all declarations with names that match a key in the map in order
  to extract their comments

This also adds a Make target for auto-generating docs partials. We can
use this for any additional docs generation.