Speed up the proxy's TDP connection
The proxy currently has to decode and re-encode each PNG frame before
sending it to the browser, because the PNG frame TDP message lacks
an explicit length field.

Introduce a new TDP message: "PNG Frame 2"

This message is identical, with the exception of a length field at
the beginning. This allows decoders that don't care about the actual
PNG data (like Teleport's proxy) to skip over the encoded data without
decoding it. Benchmarks show a 97% performance improvement for each
PNG frame received.

In order to support this workflow, tdp.Conn is updated to support a
"ParseOnly" mode. It returns the message in its encoded form, ready
to be written to the websocket.

Lastly, remove the need to read a byte and unread a byte. We were
effectively checking that the TDP message ID is valid twice for
each message.